requirements.txt
streamlit>=1.30
openai>=1.0
pdfplumber>=0.10
scikit-learn>=1.3
pandas>=2.0
numpy>=1.24
python-dotenv>=1.0
rapidfuzz>=3.5
pytest>=7.4
